Simplify how list of components is exposed to 11ty
No need for functions or computing multiple times, that's what data is for!
