+++ docs/_data/components.js
@@ -1,0 +1,3 @@
+import componentList from './componentList.js';
+
+export default Object.fromEntries(componentList.map(component => [component.tagName.replace(/^wa-/, ''), component]));
